# internal/routers/api/v1/tag.go
package v1

import (
	"blog/global"
	"blog/internal/service"
	"blog/pkg/app"
	"blog/pkg/convert"
	"blog/pkg/errcode"
	"github.com/gin-gonic/gin"
)

type Tag struct {}

func NewTag()Tag{
	return Tag{}
}

func (t Tag) Get(c *gin.Context) {}
//List
// @Summary 获取多个标签
// @Produce  json
// @Param name query string false "标签名称" maxlength(100)
// @Param state query int false "状态" Enums(0, 1) default(1)
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Success 200 {object} model.Tag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags [get]
func (t Tag)List(c *gin.Context){
	param:= service.TagListRequest{}    //定义TagList的请求体
	response:=app.NewResponse(c)     //实体化响应
	valid ,errs:=app.BindAndValid(c,&param)   //为请求体绑定参数，如果发生错误，Translation中间件就会翻译
	if !valid{
		global.Logger.Infof("app.BindAndValid errs: %v",errs)  //参数绑定失败就打印日志
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))   //响应错误信息
		return
	}

	svc:=service.New(c.Request.Context())
	paper:=app.Pager{Page: app.GetPage(c),PageSize: app.GetPageSize(c)}   //将页数信息写入paper结构体
	totalRows,err:=svc.CountTag(&service.CountTagRequest{Name: param.Name,State: param.State})//获取Tag数量
	if err!=nil{
		global.Logger.Infof("svc.CountTag err:%v",err)   //发生错误就写入日志
		response.ToErrorResponse(errcode.ErrorCountTagFail)
		return
	}

	tags,err:=svc.GetTagList(&param,&paper)  //获取Tag列表
	if err!=nil{        //发生错误就写入日志
		global.Logger.Infof("svc.GetTagList err: %v", err)
		response.ToErrorResponse(errcode.ErrorCountTagFail)
		return
	}
	response.ToResponseList(tags,totalRows)
	return
}
//Create
// @Summary 新增标签
// @Produce  json
// @Param name body string true "标签名称" minlength(3) maxlength(100)
// @Param state body int false "状态" Enums(0, 1) default(1)
// @Param created_by body string true "创建者" minlength(3) maxlength(100)
// @Success 200 {object} model.Tag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags [post]
func (t Tag)Create(c *gin.Context){
	param := service.CreateTagRequest{}
	response := app.NewResponse(c)
	valid, errs := app.BindAndValid(c, &param)
	if !valid {
		global.Logger.Infof("app.BindAndValid errs: %v",errs)  //参数绑定失败就打印日志
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))   //响应错误信息
		return
	}
	svc := service.New(c.Request.Context())
	err := svc.CreateTag(&param)
	if err != nil {
		global.Logger.Infof("svc.CreateTag err: %v", err)
		response.ToErrorResponse(errcode.ErrorCreateTagFail)
		return
	}
	response.ToResponse(gin.H{})
	return
}
//Update
// @Summary 更新标签
// @Produce  json
// @Param id path int true "标签ID"
// @Param name body string false "标签名称" minlength(3) maxlength(100)
// @Param state body int false "状态" Enums(0, 1) default(1)
// @Param modified_by body string true "修改者" minlength(3) maxlength(100)
// @Success 200 {array} model.Tag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags/{id} [put]
func (t Tag)Update(c *gin.Context){
	param := service.UpdateTagRequest{
		ID: convert.StrTo(c.Param("id")).MustUInt32(),
	}
	response := app.NewResponse(c)
	valid, errs := app.BindAndValid(c, &param)
	if !valid {
		global.Logger.Infof( "app.BindAndValid errs: %v", errs)
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))
		return
	}

	svc := service.New(c.Request.Context())
	err := svc.UpdateTag(&param)
	if err != nil {
		global.Logger.Infof( "svc.UpdateTag err: %v", err)
		response.ToErrorResponse(errcode.ErrorUpdateTagFail)
		return
	}

	response.ToResponse(gin.H{})
	return
}
//Delete
// @Summary 删除标签
// @Produce  json
// @Param id path int true "标签 ID"
// @Success 200 {string} string "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags/{id} [delete]
func (t Tag)Delete(c *gin.Context){
	param := service.DeleteTagRequest{ID: convert.StrTo(c.Param("id")).MustUInt32()}
	response := app.NewResponse(c)
	valid, errs := app.BindAndValid(c, &param)
	if !valid {
		global.Logger.Infof( "app.BindAndValid errs: %v", errs)
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))
		return
	}

	svc := service.New(c.Request.Context())
	err := svc.DeleteTag(&param)
	if err != nil {
		global.Logger.Infof( "svc.DeleteTag err: %v", err)
		response.ToErrorResponse(errcode.ErrorDeleteTagFail)
		return
	}

	response.ToResponse(gin.H{})
}